--- a/systems/primitives/constant_value_source-inl.h
+++ b/systems/primitives/constant_value_source-inl.h
@@ -1,44 +1,6 @@
 #pragma once
 
-/// @file
-/// Template method implementations for constant_value_source.h.
-/// Most users should only include that file, not this one.
-/// For background, see https://drake.mit.edu/cxx_inl.html.
-
-/* clang-format off to disable clang-format-includes */
 #include "drake/systems/primitives/constant_value_source.h"
-/* clang-format on */
 
-#include <memory>
-#include <utility>
-
-namespace drake {
-namespace systems {
-
-template <typename T>
-ConstantValueSource<T>::ConstantValueSource(const AbstractValue& value)
-    : LeafSystem<T>(SystemTypeTag<systems::ConstantValueSource>{}),
-      source_value_(value.Clone()) {
-  // Use the "advanced" method to provide explicit non-member functors here
-  // since we already have AbstractValues.
-  this->DeclareAbstractOutputPort(
-      [this]() {
-        return source_value_->Clone();
-      },
-      [this](const Context<T>&, AbstractValue* output) {
-        output->SetFrom(*source_value_);
-      });
-}
-
-template <typename T>
-ConstantValueSource<T>::ConstantValueSource(
-    std::unique_ptr<AbstractValue> value)
-    : ConstantValueSource<T>(*value) {}
-
-template <typename T>
-template <typename U>
-ConstantValueSource<T>::ConstantValueSource(const ConstantValueSource<U>& other)
-    : ConstantValueSource<T>(*other.source_value_) {}
-
-}  // namespace systems
-}  // namespace drake
+// TODO(jwnimmer-tri) Remove this header after 2019-04-01.
+#warning This header is deprecated; use constant_value_source.h instead.