--- a/systems/primitives/constant_value_source.cc
+++ b/systems/primitives/constant_value_source.cc
@@ -1,7 +1,4 @@
-// NOLINTNEXTLINE(build/include) False positive on inl file.
-#include "drake/systems/primitives/constant_value_source-inl.h"
-
-#include "drake/common/default_scalars.h"
+#include "drake/systems/primitives/constant_value_source.h"
 
 DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_SCALARS(
     class ::drake::systems::ConstantValueSource)