--- a/systems/primitives/constant_value_source.h
+++ b/systems/primitives/constant_value_source.h
@@ -1,7 +1,9 @@
 #pragma once
 
 #include <memory>
+#include <utility>
 
+#include "drake/common/default_scalars.h"
 #include "drake/common/drake_copyable.h"
 #include "drake/common/drake_deprecated.h"
 #include "drake/systems/framework/leaf_system.h"
@@ -46,5 +48,33 @@ class ConstantValueSource : public LeafSystem<T> {
   const std::unique_ptr<AbstractValue> source_value_;
 };
 
+template <typename T>
+ConstantValueSource<T>::ConstantValueSource(const AbstractValue& value)
+    : LeafSystem<T>(SystemTypeTag<systems::ConstantValueSource>{}),
+      source_value_(value.Clone()) {
+  // Use the "advanced" method to provide explicit non-member functors here
+  // since we already have AbstractValues.
+  this->DeclareAbstractOutputPort(
+      [this]() {
+        return source_value_->Clone();
+      },
+      [this](const Context<T>&, AbstractValue* output) {
+        output->SetFrom(*source_value_);
+      });
+}
+
+template <typename T>
+ConstantValueSource<T>::ConstantValueSource(
+    std::unique_ptr<AbstractValue> value)
+    : ConstantValueSource<T>(*value) {}
+
+template <typename T>
+template <typename U>
+ConstantValueSource<T>::ConstantValueSource(const ConstantValueSource<U>& other)
+    : ConstantValueSource<T>(*other.source_value_) {}
+
 }  // namespace systems
 }  // namespace drake
+
+DRAKE_DECLARE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_SCALARS(
+    class ::drake::systems::ConstantValueSource)