--- a/systems/primitives/pass_through.cc
+++ b/systems/primitives/pass_through.cc
@@ -1,7 +1,4 @@
-// NOLINTNEXTLINE(build/include) False positive on inl file.
-#include "drake/systems/primitives/pass_through-inl.h"
-
-#include "drake/common/default_scalars.h"
+#include "drake/systems/primitives/pass_through.h"
 
 DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_SCALARS(
     class ::drake::systems::PassThrough)