--- a/systems/primitives/zero_order_hold.cc
+++ b/systems/primitives/zero_order_hold.cc
@@ -1,7 +1,4 @@
-// NOLINTNEXTLINE(build/include) False positive on inl file.
-#include "drake/systems/primitives/zero_order_hold-inl.h"
-
-#include "drake/common/default_scalars.h"
+#include "drake/systems/primitives/zero_order_hold.h"
 
 DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_SCALARS(
     class ::drake::systems::ZeroOrderHold)