We should be consistent within the `primitives` package about how we approach writing a system and supporting different scalars.  My proposal is that we put the methods in the (same) header, just like we do with `framework`.

Most all changes here are just relocating code.  The only other meaningful changes are
(1) adding deprecation messages, and
(2) adding `DRAKE_DECLARE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_SCALARS` for these headers.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/robotlocomotion/drake/10432)
<!-- Reviewable:end -->