@jwnimmer-tri 2019-01-18
+@sherm1 for all review, please.

Using the following test script, the build performance when editing these files and doing an incremental rebuild is:
- on master: 5 process, 20 second wait time
- with the PR: 20 processes, 40 second wait time

It doesn't seem like moving the method bodies leads to much of a disruption at all.

The test script is an incremental rebuild after changing _all four_ of these components at once, rebuilding _all_ of the inside-Drake, non-Python code that uses them.

```bash
#!/bin/bash

set -ex

# Switch to master; see what depends on the components we care about.
git checkout master
bazel query '//... intersect allpaths(kind(cc_.*, //... except //bindings/... except //attic/... except //tools/install/libdrake/...), set(//systems/primitives:constant_value_source //systems/primitives:gain //systems/primitives:sine //systems/primitives:zero_order_hold))' | sort | tee measure-packages.txt

# Refresh everything.
bazel build $(cat measure-packages.txt)

# Change the PR files then rebuild.
echo "const int dummy1 = 1;" >> systems/primitives/constant_value_source-inl.h
echo "const int dummy2 = 2;" >> systems/primitives/gain-inl.h
echo "const int dummy3 = 3;" >> systems/primitives/sine-inl.h
echo "const int dummy4 = 4;" >> systems/primitives/zero_order_hold-inl.h
time bazel build $(cat measure-packages.txt)

# Switch to the PR.
git checkout -- systems/primitives/constant_value_source-inl.h systems/primitives/gain-inl.h systems/primitives/sine-inl.h systems/primitives/zero_order_hold-inl.h
git checkout inl

# Refresh everything.
bazel build $(cat measure-packages.txt)

# Change the PR files then rebuild.
echo "const int dummy1 = 1;" >> systems/primitives/constant_value_source.h
echo "const int dummy2 = 2;" >> systems/primitives/gain.h
echo "const int dummy3 = 3;" >> systems/primitives/sine.h
echo "const int dummy4 = 4;" >> systems/primitives/zero_order_hold.h
time bazel build $(cat measure-packages.txt)

# All done.
git checkout -- systems/primitives/constant_value_source.h systems/primitives/gain.h systems/primitives/sine.h systems/primitives/zero_order_hold.h
git checkout master
```