@sherm1 2019-01-18
For a more apples-to-apples compilation test, I changed the script to modify both the .h and -inl.h files in the master branch (since together those have the same content as the .h files in the pr). I also removed diagram_test. That gave 19 processes for both compilations, with these timings on my Puget with nothing else happening:
```
master: 
  33.8s (critical path 32.6s)
  34.7 (critical path 32.7)
  34.2 (critical path 33.1)
pr: 
  35.8 (critical path 35.0s)
  32.6 (critical path 31.8s)
  33.2 (critical path 32.6s)
```
There appears to be not significant effect on compilation time. That's a win!